SM-1287: Refactor command module. Refactor project list out of main.rs
