--- crates/bws/src/command/config.rs
@@ -1,42 +1,0 @@
-use std::{path::PathBuf, str::FromStr};
-
-use bitwarden::auth::AccessToken;
-use color_eyre::eyre::{bail, Result};
-
-use crate::{config, ProfileKey};
-
-pub(crate) fn execute(
-    name: Option<ProfileKey>,
-    value: Option<String>,
-    delete: bool,
-    profile: Option<String>,
-    access_token: Option<String>,
-    config_file: Option<PathBuf>,
-) -> Result<()> {
-    let profile = if let Some(profile) = profile {
-        profile
-    } else if let Some(access_token) = access_token {
-        AccessToken::from_str(&access_token)?
-            .access_token_id
-            .to_string()
-    } else {
-        String::from("default")
-    };
-
-    if delete {
-        config::delete_profile(config_file.as_deref(), profile)?;
-        println!("Profile deleted successfully!");
-    } else {
-        let (name, value) = match (name, value) {
-            (None, None) => bail!("Missing `name` and `value`"),
-            (None, Some(_)) => bail!("Missing `value`"),
-            (Some(_), None) => bail!("Missing `name`"),
-            (Some(name), Some(value)) => (name, value),
-        };
-
-        config::update_profile(config_file.as_deref(), profile, name, value)?;
-        println!("Profile updated successfully!");
-    };
-
-    Ok(())
-}
